0.3.0 Commit
Removed logic for delta updating in the APA102 driver. Since the users
of the driver might have another source for output data, it might be
more efficient to compare the data in its native format instead of
having the driver check for changes after it has been loaded into the
driver's own framebuffer, as that incurs a bunch of data processing,
lookups, and function calls, that are not necessarily implemented in C.

Added a call to close() for the gpio chip handle in the close() method
of the APA102 driver, as added in the latest lipgpiod update.

Minor abbreviation correction from wrgb to brgb, since each APA102
control frame contains brightness + RGB LED intensity instead of White
LED intensity + RGB LED intensity.

Removed delta update test.